edge/requirements.txt
numpy>=1.24,<2.0
scipy>=1.11,<2.0
scikit-learn>=1.3,<2.0
pyzmq>=25.0,<27.0
PyYAML>=6.0,<7.0
